requirements.txt
pyvmomi==9.0.0.0
PyYAML==6.0.2
logging==0.4.9.6
Jinja2==3.1.6